requirements.txt
coverage
Django
djangorestframework
django-environ
mysqlclient
requests
virtualenv
pytz
boto
boto3
django-storages
